rename the gc for terminated pods to "podgc"
