+++ pkg/controller/podgc/doc.go
@@ -14,11 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package gc contains a very simple pod "garbage collector" implementation,
-// GCController, that runs in the controller manager. If the number of pods
+// Package podgc contains a very simple pod "garbage collector" implementation,
+// PodGCController, that runs in the controller manager. If the number of pods
 // in terminated phases (right now either Failed or Succeeded) surpasses a
 // configurable threshold, the controller will delete pods in terminated state
-// until the system reaches the allowed threshold again. The GCController
+// until the system reaches the allowed threshold again. The PodGCController
 // prioritizes pods to delete by sorting by creation timestamp and deleting the
-// oldest objects first. The GCController will not delete non-terminated pods.
-package gc
+// oldest objects first. The PodGCController will not delete non-terminated
+// pods.
+package podgc
